Add templates and static files to the project
index.html was added as a template to the "meetups" app, and base.css
was added as a static styling file to the "meetups" app. Instead of
returning a static message through the views, the view is now made to
return a rendered HTML page, which has been defined as mentioned above.
The HTML file will parse, find and execute the instructions, finally
returning or rendering the processed content.