Implement dynamic URLs and template inheritance
The base landing page identifier can be changed over the lifetime of an
application. Hence, it is a good idea to have that dynamici, by defining
it and naming it in the application's url.py file.

Also, there might be a chance that a large portion of the code is
written repeatedly in various html files. This can be avoided by making
use of the template inheritance functionality offered by django, out of
the box. We can make use of tags such as extends and includes to acheive
this.