Introduce models and create user for admin console
Instead of having the dummy data hard coded inside the project, a
database model is created where data can be inserted when required.
Also, admin user is created to handle and manage functions/operations on
the django admin console.

While creating a super user, the command has to be prefixed with winpty
to avoid "Superuser creation skipped due to not running in a TTY" error.