Updates in meetup details
A new model 'Participant' was added. This is for the form which will be
built later.
New fields organizer_email, date and participants have been added to the
Meetup model. This is to display correct information on the meetup
details page. Images are not formatted to also be displayed on each
meetup detail page.
New images were uploaded.
Template files are updated to accomodate changes in models, views and
admin py scripts.