User registration
The functionality to register a user to a meetup was implemented. A lot
of build in django features come in handy. The email will be checked for
proper/valid email id format and also get inserted into the database, or
the participant list in the admin console upon successful submission. A
confirmation page is also given to the user upon submitting a valid
email address.

Django is so cool!