Enable email option
The user should be able to send an email to the organizer. This is
enabled by making use of the common mailto keyword in the html template
file, which is not django specific. Enables when the details of the
meetup is opened and also when a user registers for a specific meetup.